refactor(intake): align adapter-owned wallet processing
Why:
- move wallet processing responsibility to the adapters before later fixture and package work depends on it
- reduce duplicated wallet handling logic across inventory and intake code paths

What:
- align the intake flow around adapter-owned wallet processing and split the shared wallet inventory support
- update the related docs and tests to reflect the new ownership boundary

Checks:
- not run individually (covered by branch-level validation)
