+++ 06_scripts/pipeline_common.py
@@ -221,86 +221,70 @@ def detect_csv_header(path: Path) -> tuple[list[str], int]:
 def classify_file_family(path: Path, header: Sequence[str]) -> str:
     name = path.name.lower()
     header_lower = [column.strip().lower() for column in header]
+    header_set = set(header_lower)
+
+    def has_all(*columns: str) -> bool:
+        return set(columns).issubset(header_set)
 
     if path.suffix.lower() == ".pdf":
         return "statement_balance_pdf"
-    if "coinbase pro - fills" in name:
-        return "fills_csv"
-    if "coinbase pro - statement" in name:
-        return "transfer_statement_csv"
-    if "statement - all time" in name:
+    if has_all("id", "timestamp", "transaction type") and ("asset" in header_set or "statement" in name):
         return "custodial_all_time_csv"
-    if ("activity" in name or "activities" in name) and "export" in name:
-        return "broker_activity_csv"
-    if "monthly-statement-transactions" in name:
-        return "statement_transaction_csv"
-    if "portfolio" in name:
-        return "portfolio_snapshot_csv"
-    if "export-address-token" in name:
-        return "explorer_token_transfer_csv"
-    if "export-internal-tx" in name:
-        return "explorer_internal_transaction_csv"
-    if "export-address-nfts" in name:
-        return "explorer_nft_transfer_csv"
-    if "export-" in name:
-        return "explorer_transaction_csv"
-    if "ledgerlive-operations" in name:
-        return "wallet_operation_csv"
-    if "futures-trade-history" in name or "spot-trade-history" in name:
+    if has_all("portfolio", "trade id", "product", "side", "created at"):
         return "fills_csv"
-    if "convert-order-history" in name:
-        return "convert_order_csv"
-    if "deposit-history" in name and "fiat" not in name:
-        return "deposit_history_csv"
-    if "withdraw-history" in name and "fiat" not in name:
-        return "withdrawal_history_csv"
-    if "c2c-order-history" in name:
-        return "p2p_order_csv"
-    if "fiat-buy-history" in name:
-        return "fiat_buy_csv"
-    if "fiat-sell-history" in name:
-        return "fiat_sell_csv"
-    if "fiat-exchange-history" in name:
-        return "fiat_exchange_csv"
-    if "futures-transaction-history" in name:
-        return "futures_transaction_csv"
-    if "transaction" in name and "history" in name:
-        return "custodial_transaction_csv"
-    if "cash_transactions" in name:
-        return "fiat_transaction_csv"
-    if "crypto_transactions" in name:
-        return "custodial_transaction_csv"
-    if name.endswith("my_trading_history_report.csv"):
-        return "derivatives_report_csv"
-    if "_ft_transactions_" in name:
-        return "near_ft_transaction_csv"
-    if "_nft_transactions_" in name:
-        return "near_nft_transaction_csv"
-    if "_receipts_" in name:
-        return "near_receipt_csv"
-    if "_transactions_" in name and "ft_" not in name and "nft_" not in name:
-        return "near_transaction_csv"
-    if {"date", "pair", "addr"}.issubset(set(header_lower)):
-        return "derivatives_report_csv"
-    if "receipt" in header_lower and "deposit value" in header_lower:
-        return "near_receipt_csv"
-    if "txn hash" in header_lower and "direction" in header_lower and "token id" in header_lower:
-        return "near_nft_transaction_csv"
-    if "txn hash" in header_lower and "direction" in header_lower and "token" in header_lower:
-        return "near_ft_transaction_csv"
-    if "txn hash" in header_lower and "method" in header_lower and "deposit value" in header_lower:
-        return "near_transaction_csv"
-    if "statement" in header_lower and "time" in header_lower:
+    if has_all("portfolio", "type", "time", "amount", "balance", "amount/balance unit"):
         return "transfer_statement_csv"
-    if "trade id" in header_lower and "product" in header_lower:
-        return "fills_csv"
-    if "operation type" in header_lower and "operation amount" in header_lower:
-        return "wallet_operation_csv"
-    if "transaction hash" in header_lower:
-        return "explorer_transaction_csv"
-    if "activity_type" in header_lower:
+    if has_all("transaction_date", "settlement_date", "account_type", "activity_type"):
         return "broker_activity_csv"
-    if "type" in header_lower and "amount credited" in header_lower:
+    if has_all("date", "transaction", "description", "amount", "balance", "currency"):
+        return "statement_transaction_csv"
+    if has_all("operation date", "operation type", "operation amount"):
+        return "wallet_operation_csv"
+    if has_all("date", "pair", "addr"):
+        return "derivatives_report_csv"
+    if "receipt" in header_set and "deposit value" in header_set:
+        return "near_receipt_csv"
+    if has_all("txn hash", "direction", "token id", "contract"):
+        return "near_nft_transaction_csv"
+    if has_all("txn hash", "direction", "token", "contract"):
+        return "near_ft_transaction_csv"
+    if has_all("txn hash", "method", "deposit value", "txn fee"):
+        return "near_transaction_csv"
+    if has_all("transaction hash", "blockno", "unixtimestamp", "datetime (utc)", "tokenvalue", "tokensymbol"):
+        return "explorer_token_transfer_csv"
+    if has_all("transaction hash", "blockno", "unixtimestamp", "datetime (utc)", "token id", "quantity"):
+        return "explorer_nft_transfer_csv"
+    if has_all("transaction hash", "blockno", "unixtimestamp", "datetime (utc)", "parenttxfrom", "parenttxto"):
+        return "explorer_internal_transaction_csv"
+    if "transaction hash" in header_set and "datetime (utc)" in header_set and any(
+        column.startswith("value_in(") or column.startswith("value_out(") for column in header_lower
+    ):
+        return "explorer_transaction_csv"
+    if has_all("type", "amount credited", "asset credited", "amount debited", "asset debited"):
+        return "custodial_transaction_csv"
+    if has_all("date", "type", "description", "debit", "credit"):
+        return "fiat_transaction_csv"
+    if has_all("timestamp (utc)", "transaction description", "currency", "amount", "transaction kind"):
+        return "custodial_transaction_csv"
+    if has_all("time", "wallet", "pair", "sell", "buy", "status"):
+        return "convert_order_csv"
+    if has_all("time", "coin", "network", "amount", "address", "txid", "status"):
+        return "deposit_history_csv"
+    if has_all("time", "coin", "network", "amount", "fee", "address", "txid", "status"):
+        return "withdrawal_history_csv"
+    if has_all("order number", "order type", "asset", "fiat type", "total price", "status"):
+        return "p2p_order_csv"
+    if has_all("time", "method", "spend amount", "receive amount", "fee", "price", "status", "transaction id"):
+        return "fiat_buy_csv"
+    if has_all("method", "amount", "price", "final amount", "created time", "status", "transaction id"):
+        return "fiat_exchange_csv"
+    if has_all("time", "type", "amount", "asset", "symbol", "transaction id"):
+        return "futures_transaction_csv"
+    if has_all("time", "pair", "side", "price", "executed", "amount", "fee"):
+        return "fills_csv"
+    if has_all("chain", "token", "amount", "value") or "portfolio" in name:
+        return "portfolio_snapshot_csv"
+    if "transaction" in name and "history" in name:
         return "custodial_transaction_csv"
     return "unknown"
