+++ tests/unit/test_source_adapters.py
@@ -22,15 +22,16 @@ class SourceAdapterTests(unittest.TestCase):
         self.assertTrue(source_adapters.get_adapter("Crypto.com").supported)
         self.assertEqual("shakepay", source_adapters.get_adapter("Shakepay").name)
         self.assertTrue(source_adapters.get_adapter("Shakepay").supported)
-        self.assertEqual("ledger_live", source_adapters.get_adapter("Ledger Live").name)
-        self.assertTrue(source_adapters.get_adapter("Ledger Live").supported)
         self.assertEqual("ledger_live", source_adapters.get_adapter("ledger-live-main").name)
-        self.assertEqual("near", source_adapters.get_adapter("NEAR Wallet").name)
-        self.assertTrue(source_adapters.get_adapter("NEAR Wallet").supported)
+        self.assertTrue(source_adapters.get_adapter("ledger-live-main").supported)
+        self.assertEqual("near", source_adapters.get_adapter("near-main").name)
+        self.assertTrue(source_adapters.get_adapter("near-main").supported)
         self.assertEqual("gtrade", source_adapters.get_adapter("GTrade 1CT").name)
         self.assertTrue(source_adapters.get_adapter("GTrade 1CT").supported)
         self.assertEqual("evm_explorer", source_adapters.get_adapter("bsc-metamask1").name)
         self.assertTrue(source_adapters.get_adapter("bsc-metamask1").supported)
+        self.assertEqual("ledger_live", source_adapters.get_adapter("Ledger Live").name)
+        self.assertEqual("near", source_adapters.get_adapter("NEAR Wallet").name)
 
     def test_get_adapter_can_resolve_from_profile_before_source_label(self) -> None:
         with tempfile.TemporaryDirectory() as tmpdir:
@@ -151,10 +152,10 @@ class SourceAdapterTests(unittest.TestCase):
         self.assertEqual("2022-01-04 18:19:41", result.canonical_events[0]["timestamp"])
 
     def test_ledger_live_adapter_normalizes_repo_exports(self) -> None:
-        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "ledger live" / "raw"
-        adapter = source_adapters.get_adapter("Ledger Live")
+        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "ledger-live-main" / "2026-03"
+        adapter = source_adapters.get_adapter("ledger-live-main")
         profile = pipeline_common.build_source_profile(
-            source="Ledger Live",
+            source="ledger-live-main",
             raw_dir=raw_dir,
             adapter_name=adapter.name,
             adapter_supported=adapter.supported,
@@ -170,10 +171,10 @@ class SourceAdapterTests(unittest.TestCase):
         self.assertIn("Withdrawal", kinds)
 
     def test_near_adapter_normalizes_repo_exports(self) -> None:
-        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "near" / "raw"
-        adapter = source_adapters.get_adapter("NEAR Wallet")
+        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "near-main" / "2026-03"
+        adapter = source_adapters.get_adapter("near-main")
         profile = pipeline_common.build_source_profile(
-            source="NEAR Wallet",
+            source="near-main",
             raw_dir=raw_dir,
             adapter_name=adapter.name,
             adapter_supported=adapter.supported,
@@ -183,7 +184,7 @@ class SourceAdapterTests(unittest.TestCase):
 
         self.assertEqual(10, len(result.canonical_events))
         self.assertEqual([], result.exceptions)
-        self.assertTrue(any(row["source"] == "NEAR Wallet - Staking" for row in result.canonical_events))
+        self.assertTrue(any(row["source"] == "near-main - Staking" for row in result.canonical_events))
         self.assertTrue(any(row["event_kind"] == "Airdrop" for row in result.canonical_events))
         self.assertIn(
             "2022-01-30 09:31:41",
@@ -213,7 +214,7 @@ class SourceAdapterTests(unittest.TestCase):
         )
 
     def test_evm_explorer_adapter_normalizes_bsc_repo_exports(self) -> None:
-        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "metamask" / "raw"
+        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "bsc-metamask1" / "2026-03"
         adapter = source_adapters.get_adapter("bsc-metamask1")
         profile = pipeline_common.build_source_profile(
             source="bsc-metamask1",
@@ -230,7 +231,7 @@ class SourceAdapterTests(unittest.TestCase):
         self.assertTrue(any(row["event_kind"] == "Staking" for row in result.canonical_events))
 
     def test_evm_explorer_adapter_surfaces_polygon_review_rows_without_importing_them(self) -> None:
-        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "metamask" / "raw"
+        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "polygon-metamask1" / "2026-03"
         adapter = source_adapters.get_adapter("polygon-metamask1")
         profile = pipeline_common.build_source_profile(
             source="polygon-metamask1",
@@ -247,7 +248,7 @@ class SourceAdapterTests(unittest.TestCase):
         self.assertTrue(any("suspicious NFT airdrop" in row["message"] for row in result.exceptions))
 
     def test_evm_explorer_adapter_surfaces_eth_gala_review_rows_without_importing_them(self) -> None:
-        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "metamask" / "raw"
+        raw_dir = REPO_ROOT / "01_raw_exports" / "external" / "eth-gala1" / "2026-03"
         adapter = source_adapters.get_adapter("eth-gala1")
         profile = pipeline_common.build_source_profile(
             source="eth-gala1",
@@ -369,12 +370,12 @@ class SourceAdapterTests(unittest.TestCase):
             ("Coinbase", REPO_ROOT / "01_raw_exports" / "external" / "coinbase" / "raw"),
             ("WealthSimple", REPO_ROOT / "01_raw_exports" / "external" / "wealthsimple" / "raw"),
             ("Binance", REPO_ROOT / "01_raw_exports" / "external" / "binance" / "raw"),
-            ("bsc-metamask1", REPO_ROOT / "01_raw_exports" / "external" / "metamask" / "raw"),
+            ("bsc-metamask1", REPO_ROOT / "01_raw_exports" / "external" / "bsc-metamask1" / "2026-03"),
             ("Crypto.com", REPO_ROOT / "01_raw_exports" / "external" / "crypto.com" / "raw"),
             ("GTrade 1CT", REPO_ROOT / "01_raw_exports" / "external" / "gtrade" / "raw"),
             ("Shakepay", REPO_ROOT / "01_raw_exports" / "external" / "shakepay" / "raw"),
-            ("Ledger Live", REPO_ROOT / "01_raw_exports" / "external" / "ledger live" / "raw"),
-            ("NEAR Wallet", REPO_ROOT / "01_raw_exports" / "external" / "near" / "raw"),
+            ("ledger-live-main", REPO_ROOT / "01_raw_exports" / "external" / "ledger-live-main" / "2026-03"),
+            ("near-main", REPO_ROOT / "01_raw_exports" / "external" / "near-main" / "2026-03"),
         ]
 
         for source, raw_dir in cases:
@@ -394,11 +395,11 @@ class SourceAdapterTests(unittest.TestCase):
             ("Coinbase", REPO_ROOT / "01_raw_exports" / "external" / "coinbase" / "raw"),
             ("WealthSimple", REPO_ROOT / "01_raw_exports" / "external" / "wealthsimple" / "raw"),
             ("Binance", REPO_ROOT / "01_raw_exports" / "external" / "binance" / "raw"),
-            ("bsc-metamask1", REPO_ROOT / "01_raw_exports" / "external" / "metamask" / "raw"),
+            ("bsc-metamask1", REPO_ROOT / "01_raw_exports" / "external" / "bsc-metamask1" / "2026-03"),
             ("Crypto.com", REPO_ROOT / "01_raw_exports" / "external" / "crypto.com" / "raw"),
             ("Shakepay", REPO_ROOT / "01_raw_exports" / "external" / "shakepay" / "raw"),
-            ("Ledger Live", REPO_ROOT / "01_raw_exports" / "external" / "ledger live" / "raw"),
-            ("NEAR Wallet", REPO_ROOT / "01_raw_exports" / "external" / "near" / "raw"),
+            ("ledger-live-main", REPO_ROOT / "01_raw_exports" / "external" / "ledger-live-main" / "2026-03"),
+            ("near-main", REPO_ROOT / "01_raw_exports" / "external" / "near-main" / "2026-03"),
             ("GTrade 1CT", REPO_ROOT / "01_raw_exports" / "external" / "gtrade" / "raw"),
         ]
